Fixing mismatch between MovieRepository and Movie Id class

+++ src/main/java/com/ivanmiskulin/demo/MovieRepo.java
@@ -8,7 +8,7 @@ import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 import com.ivanmiskulin.demo.entity.Movie;
 
 @RepositoryRestResource(collectionResourceRel = "movie", path = "movies")
-public interface MovieRepo extends JpaRepository<Movie, Long> {
+public interface MovieRepo extends JpaRepository<Movie, Integer> {
 
 	public List<Movie> findByTitle(String title);
 }
